refactor(cli): move cli to internal
