+++ cmd/cli/main.go
@@ -1,71 +1,12 @@
 package cli
 
 import (
+	"container-dsh/internal/cli"
 	"fmt"
 	"os"
 
 	tea "github.com/charmbracelet/bubbletea"
 )
-
-type Model struct {
-	choices  []string
-	cursor   int
-	selected map[int]struct{}
-}
-
-func (m Model) Init() tea.Cmd {
-	return nil
-}
-
-func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch msg.String() {
-
-		case "j", "down":
-			if m.cursor < len(m.choices)-1 {
-				m.cursor++
-			}
-		case "k", "up":
-			if m.cursor > 0 {
-				m.cursor--
-			}
-		case "ctrl+c", "q":
-			return m, tea.Quit
-		case "enter", " ":
-			_, ok := m.selected[m.cursor]
-			if ok {
-				delete(m.selected, m.cursor)
-			} else {
-				m.selected[m.cursor] = struct{}{} // TODO: Show the Output of the selected option
-			}
-		}
-	}
-	return m, nil
-}
-
-func (m Model) View() string {
-	s := "Select options:\n\n"
-
-	for i, choice := range m.choices {
-		cursor := " "
-		if m.cursor == i {
-			cursor = ">"
-		}
-
-		checked := " "
-		if _, ok := m.selected[i]; ok {
-			checked = "x"
-		}
-
-		s += fmt.Sprintf("%s [%s] %s\n", cursor, checked, choice)
-	}
-
-	s += "\nPress q to quit.\n"
-
-	return s
-}
 
 func Run() {
 	fmt.Println("CLI VERSION")
@@ -78,9 +19,9 @@ func Run() {
 
 }
 
-func initialModel() Model {
-	return Model{
-		choices:  []string{"Show containers", "show images", "show stats"},
-		selected: make(map[int]struct{}),
+func initialModel() cli.Model {
+	return cli.Model{
+		Choices:  []string{"Show containers", "show images", "show stats"},
+		Selected: make(map[int]struct{}),
 	}
 }
